Add CAST-Manual (preliminary version)
Added the CAST-Manual as LaTeX-Document to the repository. I know some
people will hate me for the choice of LaTeX over the existing (although
very unfinished) CAST-Manual in Microsoft-Word-Format. However, LaTeX
and biblatex files are human readable and can smoothly be tracked by git
versioning. I have transferred all of the existing content of the
msword-manual to latex. Please note that this manual is still missing
lots of stuff and is to be considered unfinished. However, I do thing I
have added most of the conceptual stuff necessary, such as the
bibliography, abbreviation support and so on. If you feel this is a bad
idea, please comment on this. I have not found a suitable non-binary
format for MSWord to store the manual in for versioning, so LaTeX seems
an obvious choice. Also note the sue of conditionals in the document for
creating different versions.