(Quick)Fix for faulty coords* in aco_ff::refined
Due to some error in the move constructor called for
tinker::refine::refined, there was a faulty coords* contained. This led
to errors on some machines when performing an MD Simulation. This is a
quickfix for this problem by means of an auxilary function. At some
point a custom move constructor should be written, fixing this issue
more beautifully. Fixes #24 .