Add check for NaNs to MD Simulation
According to some test on my local windows machine this does not impact
performance meassurably. I ran 30 tests in total using 10k and 100k
steps with and without the checks implemented and with and without
optimization. In all cases, differences were smaller than the
fluctuations between different simulations (also the mean differences).
Furthermore, this check will seldomly be a limiting factor as it doesn't
scale with system size but only with simulation time. Closes #2.