Fix critical bug in quadratic matrix transposition
This would yield the original matrix as a result of transposition of a
quadratic matrix (duh, look at tit, just a stupid mastike :D).
Consequently, all kabsch alignment prior to this commit in branch devel
was faulty (I don't know when the bug was introduced). However, branch
master still worked and works correctly.