Implement failcounter for MOPAC calls
If a MOPAC call fails, a failcounter will be incremented and an error
will be displayed to cerr (independent of verbosity). If failcounter
gets bigger than 1000, CAST will abort. Manual entry discribing this
behavior has been added. Failcounter is part of energy base class for
simplicity's sake, it was easier to code since no move / etc calls had
to be rewritten. Failcounter should also be implemented for TeraChem MPI
calls in the future. Fixes #6. Fixes #31.