Assign tinker FF types from amber.prm when reading input from amber files
This is sort of a good example of shooting yourself in the foot. I have
been thinking for a while baout how to implement the IO of AMBER FF
types painlessly. The easiest way is actually to just asign random atom
types that correspond to the correct TINKER atom classes. When the FF
parameters are then contracted accoring to the prm-file and amber.prm is
provided, the correct types will be yielded. Also moved AMBER IO to its
own file since it is getting lengthy. Closes #7.