Remove deprecated trace_offset from MD and add MDtrackoffset
Probably the currently not used "trace_offset" once had the same
function as the new trackoffset: Controlling every nth frame on which
the TRACE output is written (probably could have kept the name now that
I think about it.... oh well). This also removes a nasty bug where the
MD pressure was not initialized and sometimes (only sometimes)
uninitialized numbers would be printed and used which was quite
confusing.